refactor(adapters): adapter factory を composables から adapters に移動
useInitAdapter は composable ではなく単なる async ファクトリ関数なので、
use* 命名違反を解消して adapters/factory.ts に移動。
adapter のライフサイクル (cache/destroy) を adapter 層に集約。
