+++ src/components/window/FollowListContent.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { computed, onMounted, ref, watch } from 'vue'
+import { initAdapterFor } from '@/adapters/factory'
 import type {
   FollowRelation,
   NormalizedUser,
@@ -8,7 +9,6 @@ import type {
 import LoadingSpinner from '@/components/common/LoadingSpinner.vue'
 import MkAvatar from '@/components/common/MkAvatar.vue'
 import MkMfm from '@/components/common/MkMfm.vue'
-import { initAdapterFor } from '@/composables/useInitAdapter'
 import { showLoginPrompt } from '@/composables/useLoginPrompt'
 import { useNavigation } from '@/composables/useNavigation'
 import { isGuestAccount, useAccountsStore } from '@/stores/accounts'
