+++ src/stores/accounts.ts
@@ -2,8 +2,8 @@ import { listen } from '@tauri-apps/api/event'
 import JSON5 from 'json5'
 import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
+import { destroyAdapter } from '@/adapters/factory'
 import type { ServerSoftware } from '@/adapters/types'
-import { destroyAdapter } from '@/composables/useInitAdapter'
 import { deleteAllMemos } from '@/composables/useMemos'
 import { readAccountOrder, writeAccountOrder } from '@/utils/settingsFs'
 import { removeStorage, STORAGE_KEYS } from '@/utils/storage'
